Remote Templates (#329)
* install `gittar`

* remove flags: css,sass,less,dest,type

* purge *css flag-related stuffs

* extract yarn-vs-npm logic before printing

* replace ansi-codes with `chalk.green`

* extract `pkgFile` path

* use `argv.name` within pkg & manifest (Closes #326)

- declare `log-symbols` as a dependency
  - it is already a direct dependency of `ora`

* extract an `error` helper

- with nice symbol & formatting

* uninstall `recursive-copy`

* extract `isDir` helper

- uses *sync versions, but oh well

* replace `copy` with `gittar.extract` - with filter

* move `target` + `exists` + `force` logic first;

- will need to verify a destination before extracting/downloading
- will prompt (TODO) if incomplete
- will abort if unsafe

* use `argv.dest` for final instructions

* fix typo

* export `setup` functions directly

* remove debug-type stdout from `initGit`

* move common helpers to `util` file;

- use more partial import statements

* move `commandExists` —> `util.hasCommand`;

- uses which.sync 🙊
- update function calls
- remove `async` from `pkgScripts`

* check if `yarn` bin exists on flag parsing;

- instead of checking 5x per `setup` function
- return `async` function directly from `initialize`

* consolidate `setup.pkgScripts` function

* do not write a `gitignore` file

- template concern

* minor cleanup, warn if no `git` binary found

* oops, fix regex tester

* don’t assume exit within `util.error`

* use `warn` within watch command

* use `util.isDir` within serve command

* remove `init` command publicly

* Merge `master` into `templates`

* alias “official” preact-cli templates by name

- only `default` & `full` for now

* simplify `install` — install template deps

* move `trimLeft` —> `util.trim`

* remove `initialize` from setup;

- assumed template ships own `package.json` already

* simplify package-related `create`

- change `pkg.name` & `mani.name` if set
- set default `pkg.scripts` if none provided
   -  will dynamically install `if-env` for `start` script

* only extract `/template/` filepaths

- strip:2 to remove “template” from dest name

* add extra TODO note

* define `done` handler before export

* update `test/subjects` names

- fixes test? /shrug

* [MAJOR] tests refactor

- supports new `create` command
- refactor utility methods
- change snapshot handling
  — comparison & taking
- rename `WITH_INSTALL` —> `SKIP_INSTALL`
- remove `setup` step; not needed

* move `serve.snapshot` —>`images/serve`

* add `deprecated` notice on `build.snapshot` file

* remove unused `exists` helper

* Fix tests/image/create.js test error

* Revert "Fix tests/image/create.js test error"

This reverts commit ca3b2355da71adcc99fe52b1d3599b8b9df87786.

* [Travis] only build `master` branch

- PRs are still built

* remove `commands/init` file

- somehow `recursive-copy` was picked up?

* add console statement for debug

- WTF HAPPENING

* remove `dependency-install-loader`

- no need anymore!
- templates responsible own dependecies
- and style-* flags removed

* update scripts path

- OCD strikes again

* exit early if run `build` w/o installing deps;

- tell user to run `npm install` first

* always install deps for build subjects

* remove `SKIP_INSTALL` global for Travis (temp)

* assume any template without `/` is official lookup

- print `info` message about the assumption

* update tests, remove `full` alias

+++ src/commands/build.js
@@ -1,5 +1,6 @@
-import { resolve } from 'path';
 import rimraf from 'rimraf';
+import { resolve } from 'path';
+import { isDir, error } from '../util';
 import asyncCommand from '../lib/async-command';
 import runWebpack, { showStats, writeJsonStats } from '../lib/webpack/run-webpack';
 
@@ -48,8 +49,15 @@ export default asyncCommand({
 	},
 
 	async handler(argv) {
+		let cwd = argv.cwd ? resolve(argv.cwd) : process.cwd();
+		let modules = resolve(cwd, 'node_modules');
+
+		if (!isDir(modules)) {
+			return error('No `node_modules` found! Please run `npm install` before continuing.', 1);
+		}
+
 		if (argv.clean) {
-			let dest = resolve(argv.cwd || process.cwd(), argv.dest || 'build');
+			let dest = resolve(cwd, argv.dest || 'build');
 			await Promise.promisify(rimraf)(dest);
 		}
